🏗/🚧 separando datePicker + implementando método do service para estatísticas por local
