+++ src/utils/StatisticsUtils.js
@@ -1,0 +1,17 @@
+export const UnityTimes = {
+  DAILY: 'DAILY',
+  WEEKLY: 'WEEKLY',
+  MONTHLY: 'MONTHLY',
+};
+
+export const DateParameters = {
+  START: 0,
+  END: 1,
+};
+
+export const StatisticsType = {
+  TIME: 'time',
+  PRODUCT: 'product',
+  CATEGORY: 'category',
+  PURCHASE_LOCAL: 'purchaseLocal',
+};
